auth-service/pyproject.toml
[tool.poetry]
name = "edu-auth-service"
version = "0.1.0"
description = "auth service for education platform"
authors = ["y"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
SQLAlchemy = "^2.0.30"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
alembic = "^1.13.1"
pydantic-settings = "^2.3.3"
uvicorn = "^0.30.1"
bcrypt = {extras = ["passlib"], version = "^4.1.3"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
